requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
urllib3==2.2.0
